feat: throws an exception if common user try to delete other users

+++ src/main/java/com/kauadev/resume_builder_app/services/UserService.java
@@ -3,11 +3,15 @@ package com.kauadev.resume_builder_app.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.kauadev.resume_builder_app.domain.user.User;
 import com.kauadev.resume_builder_app.domain.user.UserDTO;
+import com.kauadev.resume_builder_app.domain.user.exceptions.CommonUserCanNotDeleteUsersException;
 import com.kauadev.resume_builder_app.domain.user.exceptions.UserNotFoundException;
 import com.kauadev.resume_builder_app.repositories.UserRepository;
 
@@ -20,6 +24,14 @@ public class UserService {
     private PasswordEncoder passwordEncoder;
 
     // evitando o uso de this se for desnecessário! melhora a legibilidade
+
+    private User getAuthenticationPrincipalUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = (User) authentication.getPrincipal();
+
+        return user;
+    }
 
     public List<User> getAllUsers() {
         List<User> users = userRepository.findAll();
@@ -49,6 +61,11 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+        User loggedUser = getAuthenticationPrincipalUser();
+        if (!loggedUser.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            throw new CommonUserCanNotDeleteUsersException();
+        }
+
         User user = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
 
         userRepository.delete(user);
